Task 1: Components for shipping.rake bizz logic
OrderFinder locates orders on a given date. The search can be filtered
by state also.

DeliveriesCreator creates necesarry deliveries according to the
max_deliveries set on the order type. Deliveries are created under an
atomic transaction.

OrderCompleter has an private #complete? method to determine if the
order is complete according to a given rule. If the rule is satisfied
then it sets the status to complete.